Initial import;  * Funcionality for building Intents from outside the android platform that could be used when whithin the platform;  * The various communicators are now able to pass the AtmosphereIntent from the client to the device;  * The Atmosphere IME is now triggered over intent broadcasting from the Atmosphere service, rather than a shell command;  * The Atmosphere IME is now better structured;
Changeset:
 * Refactoring to the AtmosphereIME;
 * Refactoring to the inputText method of Device and UIElement classes;
 * Moddifyed all classes and methods related to the passing of information from the client to device;
 * Implementation of AtmosphereIntent that could easily be used to craete android Intents from outside the platform;
 * Implementation of Pack that can hold various types of values with String key mapping;
 * Added AndroidComponentUsageExecption that is thrown when calling methods from the android sdk , outside of the android platform;

[https://inttools.musala.com/reviewboard/r/629 $629] see #4435

git-svn-id: https://trac.musala.com/ARDC/repo@11540 34708651-a5da-c34d-b20c-10ddab38317d

+++ src/com/musala/atmosphere/client/UiElement.java
@@ -405,7 +405,7 @@ public class UiElement {
      *        - boolean indicating if the element should be revalidated prior to text inputting.
      * @return <code>true</code> if the text input is successful, <code>false</code> if it fails.
      */
-    public boolean inputText(String text, int intervalInMs) {
+    public boolean inputText(String text, long intervalInMs) {
         innerRevalidation();
         focus();
         boolean success = onDevice.inputText(text, intervalInMs);
